+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 * adding in preseason/postseason schedules
   * also play in for nba
 * added in depthchart api endpoints
+* added in current_week/todays_events across schedule/week classes
+* event has a today indicator
+* added in indicators in week/events if it is today/this week
+* added in current_week property in schedule class
+* added in events_today property in weeks class
 * 
 
 ## 0.3.3